Internalize the abstract Output class
This is a sibling to commit e3624d17f01952.
