+++ lib/IO/KafkaOutput.hpp
@@ -2,7 +2,6 @@
 #define KAFKAOUTPUT_HPP
 
 #include "turboevents-internal.hpp"
-#include "turboevents.hpp"
 #include <string>
 
 namespace TurboEvents {
