+++ lib/turboevents-internal.hpp
@@ -8,7 +8,25 @@
 namespace TurboEvents {
 
 class EventStream;
-class Output;
+struct Event;
+
+/// A class encapsulating an output destination
+class Output {
+public:
+  /// Virtual destructor
+  virtual ~Output() = 0;
+
+  /// Virtual function to make an event with a string payload
+  virtual Event *makeEvent(std::chrono::system_clock::time_point,
+                           std::string data) = 0;
+
+  /// Virtual function to make an event with an int payload
+  virtual Event *makeEvent(std::chrono::system_clock::time_point, int data) = 0;
+
+protected:
+  /// Common error handling function
+  void unimp(std::string className, std::string typeName);
+};
 
 /// A class encapsulating an input, such as a file
 class Input {
