+++ src/main.cpp
@@ -1,6 +1,7 @@
 #include "turboevents.hpp"
 
 #include <gflags/gflags.h>
+#include <iostream>
 
 DEFINE_string(output, "print", "what kind of events to produce");
 
@@ -10,14 +11,21 @@ int main(int argc, char **argv) {
   gflags::ParseCommandLineFlags(&argc, &argv, true);
 
   auto turbo = TurboEvents::TurboEvents::create();
+  if (FLAGS_output == "print")
+    turbo->setPrintOutput();
+  else if (FLAGS_output == "kafka")
+    turbo->setKafkaOutput();
+  else {
+    std::cerr << "Unknown output: " << FLAGS_output << "\n";
+    exit(1);
+  }
 
-  auto output = TurboEvents::TurboEvents::createOutput(FLAGS_output);
   for (int i = 1; i < argc; ++i) turbo->createXMLFileInput(argv[i]);
 
   turbo->createStreamInput(5);
   turbo->createStreamInput(2, 1500);
 
-  turbo->run(*output);
+  turbo->run();
 
   gflags::ShutDownCommandLineFlags();
   return 0;
